+++ packages/stream_chat/lib/src/core/models/event.dart
@@ -16,6 +16,8 @@ class Event {
     this.me,
     this.user,
     this.message,
+    this.poll,
+    this.pollVote,
     this.totalUnreadCount,
     this.unreadChannels,
     this.reaction,
@@ -67,6 +69,12 @@ class Event {
 
   /// The message sent with the event
   final Message? message;
+
+  /// The poll sent with the event
+  final Poll? poll;
+
+  /// The poll vote sent with the event
+  final PollVote? pollVote;
 
   /// The channel sent with the event
   final EventChannel? channel;
@@ -138,6 +146,8 @@ class Event {
     'me',
     'user',
     'message',
+    'poll',
+    'poll_vote',
     'total_unread_count',
     'unread_channels',
     'reaction',
@@ -170,6 +180,8 @@ class Event {
     OwnUser? me,
     User? user,
     Message? message,
+    Poll? poll,
+    PollVote? pollVote,
     EventChannel? channel,
     Member? member,
     Reaction? reaction,
@@ -191,6 +203,8 @@ class Event {
         me: me ?? this.me,
         user: user ?? this.user,
         message: message ?? this.message,
+        poll: poll ?? this.poll,
+        pollVote: pollVote ?? this.pollVote,
         totalUnreadCount: totalUnreadCount ?? this.totalUnreadCount,
         unreadChannels: unreadChannels ?? this.unreadChannels,
         reaction: reaction ?? this.reaction,