+++ packages/stream_chat/test/src/mocks.dart
@@ -10,6 +10,7 @@ import 'package:stream_chat/src/core/api/general_api.dart';
 import 'package:stream_chat/src/core/api/guest_api.dart';
 import 'package:stream_chat/src/core/api/message_api.dart';
 import 'package:stream_chat/src/core/api/moderation_api.dart';
+import 'package:stream_chat/src/core/api/polls_api.dart';
 import 'package:stream_chat/src/core/api/user_api.dart';
 import 'package:stream_chat/src/core/http/connection_id_manager.dart';
 import 'package:stream_chat/src/core/http/stream_http_client.dart';
@@ -51,6 +52,8 @@ class MockUserApi extends Mock implements UserApi {}
 class MockGuestApi extends Mock implements GuestApi {}
 
 class MockMessageApi extends Mock implements MessageApi {}
+
+class MockPollsApi extends Mock implements PollsApi {}
 
 class MockChannelApi extends Mock implements ChannelApi {}